requirements.txt
annotated-types==0.7.0
anyio==4.4.0
certifi==2024.6.2
distro==1.9.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
openai==1.33.0
pydantic==2.7.3
pydantic_core==2.18.4
pylatexenc==2.10
pynput==1.7.7
pyobjc-core==10.3
pyobjc-framework-ApplicationServices==10.3
pyobjc-framework-Cocoa==10.3
pyobjc-framework-CoreText==10.3
pyobjc-framework-Quartz==10.3
pyperclip==1.8.2
python-dotenv==1.0.1
six==1.16.0
sniffio==1.3.1
tqdm==4.66.4
typing_extensions==4.12.2
